fix: keep generic SIMD helpers in gathers
