+++ Cargo.toml
@@ -4,6 +4,7 @@ exclude = ["python"]
 resolver = "3"
 
 [workspace.dependencies]
+bytemuck = "1.22.0"
 faer = { version = "0.22.6", default-features = false, features = ["linalg", "std"] }
 pulp = "0.21.5"
 rand = "0.9.3"
@@ -26,6 +27,7 @@ rust-version = "1.85"
 [dependencies]
 aligned-vec = "0.6.4"
 argh = "0.1.13"
+bytemuck.workspace = true
 faer.workspace = true
 log = "0.4.27"
 logforth = "0.24.0"
